What a value is, as against what reads it
Four shaders: water refraction, a grass shader on two sine waves, half float
packing, and a nine tap blur. The water and the blur were clean; the other two
found five things.

A cosine of half an angle kept its range reduction. The template that takes
the reduction off shader model 3's sincos insisted the inner factor be exactly
one over two pi, and cos(x * 0.5) reduces with half of that - in the same mad
as the sin(x) beside it. The factor is the scale fxc folded in.

An integer result written to a float output is a float's bit pattern, not the
number the bits add up to. The destination storage rule covers temps; an
output has to ask its signature, since asking the instructions gets the answer
from the very thing being told apart.

An immediate two moves from the instruction that types it was typed a float,
and 8 as a float is a denormal that prints as 0. A move carries the value on,
so the type comes from what reads the register it lands in. Underneath it, the
recursion guard is the path and not everything seen, and a component the
reader overwrites is taken away only after that reader's own reads count -
`movc r1.xy, c, r1.wz, r1.xy` reads r1.y and writes it.

`(x & 0x7f800000) == 0x7f800000` came out without the brackets, and & binds
looser than ==, so it read as x & (a == b).

And a float read as an integer was converted rather than reinterpreted. Where
a value's readers agree it is an integer the variable is typed by them, but
the value can be a float whose bits are that integer, and `int t = f` rounds a
bit pattern to the number nearest it. A bitwise operator, a shift, and an
assignment into such a variable all take asint of it now.

half_packing is an aspirational fixture for what is left: where the readers
disagree the variable is one type and the other uses of it do not compile.
Recorded as a known failure, with the instruction writer's own output on it as
a fifth instance of the denormal fold.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/sample_variants.fx
@@ -14,7 +14,7 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	float4 r3;
 	r0.xy = texcoord.xy * float2(256, 256);
 	r0.xy = (int2)r0.xy;
-	r0.zw = float2(0, 0);
+	r0.zw = int2(0, 0);
 	r0 = tex.Load(r0.xyz);
 	r1 = tex.SampleLevel(samp, texcoord.xy, lod);
 	r2 = tex.SampleBias(samp, texcoord.xy, bias);
